+++ frontend/embed.go
@@ -1,0 +1,7 @@
+package frontend
+
+import "embed"
+
+// Content is a FS that holds the web UI assets
+//go:embed public/*
+var Content embed.FS