+++ server/server.go
@@ -4,8 +4,10 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"io/fs"
 	"log"
 	"net"
+	"net/http"
 	"sync"
 	"time"
 
@@ -13,9 +15,9 @@ import (
 	"github.com/nats-io/nats.go"
 	"github.com/oklog/run"
 	"github.com/simpleiot/simpleiot/api"
-	"github.com/simpleiot/simpleiot/assets/frontend"
 	"github.com/simpleiot/simpleiot/client"
 	"github.com/simpleiot/simpleiot/data"
+	"github.com/simpleiot/simpleiot/frontend"
 	"github.com/simpleiot/simpleiot/node"
 	"github.com/simpleiot/simpleiot/particle"
 	"github.com/simpleiot/simpleiot/store"
@@ -41,6 +43,7 @@ type Options struct {
 	ParticleAPIKey    string
 	AppVersion        string
 	OSVersionField    string
+	LogNats           bool
 }
 
 // Server represents a SIOT server process
@@ -48,6 +51,7 @@ type Server struct {
 	nc                 *nats.Conn
 	options            Options
 	natsServer         *server.Server
+	clients            *client.Group
 	chNatsClientClosed chan struct{}
 	chStop             chan struct{}
 	chWaitStart        chan struct{}
@@ -89,7 +93,17 @@ func NewServer(o Options) (*Server, *nats.Conn, error) {
 		chNatsClientClosed: chNatsClientClosed,
 		chStop:             make(chan struct{}),
 		chWaitStart:        make(chan struct{}),
+		clients:            client.NewGroup(),
 	}, nc, err
+}
+
+// AddClient can be used to add clients to the server.
+// Clients must be added before start is called. The
+// Server makes sure all clients are shut down before
+// shutting down the server. This makes for a cleaner
+// shutdown.
+func (s *Server) AddClient(client client.StartStop) {
+	s.clients.Add(client)
 }
 
 // Start the server -- only returns if there is an error
@@ -255,7 +269,6 @@ func (s *Server) Start() error {
 	// Build in clients manager
 	// ====================================
 
-	clientsManager := client.NewBuiltInClients(s.nc)
 	storeWg.Add(1)
 	g.Add(func() error {
 		defer storeWg.Done()
@@ -265,11 +278,11 @@ func (s *Server) Start() error {
 			return err
 		}
 
-		err = clientsManager.Start()
-		logLS("LS: Exited: clients manager")
+		err = s.clients.Start()
+		logLS("LS: Exited: clients manager: ", err)
 		return err
 	}, func(err error) {
-		clientsManager.Stop(err)
+		s.clients.Stop(err)
 		logLS("LS: Shutdown: clients manager")
 	})
 
@@ -294,14 +307,23 @@ func (s *Server) Start() error {
 		}()
 	}
 
+	// remove output dir name from frontend assets filesystem
+	feFS, err := fs.Sub(frontend.Content, "public")
+	if err != nil {
+		log.Fatal("Error getting frontend subtree: ", err)
+	}
+
+	// wrap with fs that will automatically look for and decompress gz
+	// versions of files.
+	feFSDecomp := newFsDecomp(feFS)
+
 	// ====================================
 	// HTTP API
 	// ====================================
 	httpAPI := api.NewServer(api.ServerArgs{
 		Port:       o.HTTPPort,
 		NatsWSPort: o.NatsWSPort,
-		GetAsset:   frontend.Asset,
-		Filesystem: frontend.FileSystem(),
+		Filesystem: http.FS(feFSDecomp),
 		Debug:      o.DebugHTTP,
 		JwtAuth:    auth,
 		AuthToken:  o.AuthToken,