+++ tools/tools.go
@@ -1,3 +1,4 @@
+//go:build tools
 // +build tools
 
 package tools
@@ -7,6 +8,5 @@ import (
 	// the entries from go.mod
 	// genesis is used to generate static assets
 	// to embed in binary
-	_ "github.com/benbjohnson/genesis/cmd/genesis"
 	_ "golang.org/x/lint/golint"
 )